Separate fields config logic from the content step
