+++ projects/valtimo/dossier-management/src/lib/components/dossier-management-widget-wizard/steps/widget-wizard-content-step/widget-wizard-content-step.component.html
@@ -14,61 +14,4 @@
   ~ limitations under the License.
   -->
 
-<form [formGroup]="form">
-  <cds-text-label
-    [attr.data-carbon-theme]="theme$ | async"
-    class="valtimo-widget-wizard-content__input"
-  >
-    {{ 'widgetTabManagement.content.widgetTitle' | translate }}
-
-    <input
-      formControlName="widgetTitle"
-      cdsText
-      [placeholder]="'widgetTabManagement.content.widgetTitlePlaceholder' | translate"
-      type="text"
-    />
-  </cds-text-label>
-</form>
-
-<ng-container>
-  <cds-tabs type="contained">
-    @for (column of columns(); track $index) {
-      <cds-tab
-        [heading]="columnHeading"
-        [context]="{index: $index, count: $count}"
-        (selected)="onColumnTabSelected($index)"
-      >
-      </cds-tab>
-    }
-    <cds-tab
-      *ngIf="columns().length < widgetWidth"
-      [heading]="addColumn"
-      [id]="'addColumn'"
-      class="valtimo-widget-wizard-content__add"
-    ></cds-tab>
-  </cds-tabs>
-
-  <ng-template #contentRenderer></ng-template>
-</ng-container>
-
-<ng-template #addColumn>
-  <button cdsButton="ghost" [iconOnly]="true" (click)="onAddColumnClick()">
-    <svg class="cds--btn__icon" cdsIcon="add" size="16"></svg>
-  </button>
-</ng-template>
-
-<ng-template #columnHeading let-data>
-  <div class="valtimo-widget-wizard-content__heading">
-    {{ 'widgetTabManagement.content.column' | translate: {number: data.index + 1} }}
-
-    <button
-      cdsButton="ghost"
-      [disabled]="data.count === 1"
-      [iconOnly]="true"
-      size="sm"
-      (click)="onDeleteColumnClick(index)"
-    >
-      <svg class="cds--btn__icon" cdsIcon="close" size="16"></svg>
-    </button>
-  </div>
-</ng-template>
+<ng-template #contentRenderer></ng-template>
