+++ projects/valtimo/dossier-management/src/lib/components/dossier-management-widget-wizard/steps/widget-wizard-content-step/widget-wizard-content-step.component.ts
@@ -15,28 +15,18 @@
  */
 import {CommonModule} from '@angular/common';
 import {
-  AfterViewInit,
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   Component,
   EventEmitter,
-  HostBinding,
-  OnDestroy,
   OnInit,
   Output,
   ViewChild,
   ViewContainerRef,
   ViewEncapsulation,
-  effect,
-  signal,
 } from '@angular/core';
-import {FormBuilder, FormsModule, ReactiveFormsModule, Validators} from '@angular/forms';
-import {TranslateModule} from '@ngx-translate/core';
-import {ButtonModule, IconModule, InputModule, TabsModule} from 'carbon-components-angular';
-
+import {BehaviorSubject} from 'rxjs';
 import {WidgetWizardService} from '../../../../services';
-import {CdsThemeService, CurrentCarbonTheme} from '@valtimo/components';
-import {BehaviorSubject, Subscription, combineLatest, debounceTime, map} from 'rxjs';
 
 @Component({
   selector: 'valtimo-widget-wizard-content-step',
@@ -45,99 +35,24 @@ import {BehaviorSubject, Subscription, combineLatest, debounceTime, map} from 'r
   changeDetection: ChangeDetectionStrategy.OnPush,
   encapsulation: ViewEncapsulation.None,
   standalone: true,
-  imports: [
-    CommonModule,
-    TranslateModule,
-    InputModule,
-    TabsModule,
-    IconModule,
-    ReactiveFormsModule,
-    ButtonModule,
-  ],
+  imports: [CommonModule],
 })
-export class WidgetWizardContentStepComponent implements OnDestroy, OnInit {
-  @HostBinding('class') public readonly class = 'valtimo-widget-wizard-content';
+export class WidgetWizardContentStepComponent implements OnInit {
   @ViewChild('contentRenderer', {static: true, read: ViewContainerRef})
   private readonly _vcr: ViewContainerRef;
-  @Output() public readonly validEvent = new EventEmitter<boolean>();
 
-  public form = this.fb.group({
-    widgetTitle: this.fb.control(this.widgetWizardService.widgetTitle(), Validators.required),
-  });
-
-  public readonly columns = signal<null[]>([null]);
-  public readonly widgetWidth = this.widgetWizardService.widgetWidth();
-  public readonly selectedTabIndex = -1;
-  public readonly theme$ = this.cdsThemeService.currentTheme$.pipe(
-    map((theme: CurrentCarbonTheme) =>
-      theme === CurrentCarbonTheme.G10 ? 'white' : CurrentCarbonTheme.G90
-    )
-  );
-
-  private readonly _subscriptions = new Subscription();
-  private readonly _contentValid = new BehaviorSubject<boolean>(false);
+  @Output() public validEvent = new EventEmitter<boolean>();
 
   constructor(
-    private readonly cdsThemeService: CdsThemeService,
     private readonly cdr: ChangeDetectorRef,
-    private readonly fb: FormBuilder,
     private readonly widgetWizardService: WidgetWizardService
   ) {}
 
   public ngOnInit(): void {
-    this._subscriptions.add(
-      combineLatest([this.form.valueChanges, this._contentValid])
-        .pipe(debounceTime(100))
-        .subscribe(([formValid, contentValid]) => {
-          if (formValid)
-            this.widgetWizardService.widgetTitle.set(this.form.get('widgetTitle')?.value ?? '');
-          this.validEvent.emit(formValid && contentValid);
-        })
-    );
-    const widgetContent = this.widgetWizardService.widgetContent();
-    if (!widgetContent) return;
-
-    this.columns.set(Object.keys(widgetContent).map(() => null));
+    this.renderComponent();
   }
 
-  public ngOnDestroy(): void {
-    this._subscriptions.unsubscribe();
-    this.form.reset();
-  }
-
-  public onAddColumnClick(): void {
-    this.columns.update(value => [...value, null]);
-  }
-
-  public onColumnTabSelected(index: number): void {
-    this.renderComponent(index);
-  }
-
-  public onDeleteColumnClick(index: number): void {
-    this.widgetWizardService.widgetContent.update(
-      (content: {[columnIndex: number]: any} | null) => {
-        if (!content) return null;
-
-        let tempIndex = index;
-        while (tempIndex < this.columns().length - 1) {
-          content[tempIndex] = content[tempIndex + 1];
-          tempIndex++;
-        }
-        delete content[tempIndex];
-
-        return content;
-      }
-    );
-
-    this.columns.update((columns: null[]) => {
-      const temp = columns;
-      temp.splice(index, 1);
-
-      return temp;
-    });
-  }
-
-  private renderComponent(columnIndex: number): void {
+  private renderComponent(): void {
     this._vcr.clear();
     const widget = this.widgetWizardService.selectedWidget();
     if (!widget) return;
@@ -145,18 +60,8 @@ export class WidgetWizardContentStepComponent implements OnDestroy, OnInit {
     const componentInstance = this._vcr.createComponent(widget.component).instance;
     if (!componentInstance) return;
 
-    const widgetContent = this.widgetWizardService.widgetContent();
-    if (!!widgetContent) componentInstance.columnData = widgetContent[columnIndex];
+    componentInstance.changeEvent.subscribe(event => this.validEvent.emit(event.valid));
 
-    this._subscriptions.add(
-      componentInstance.changeEvent.subscribe((event: {data: any; valid: boolean}) => {
-        this.widgetWizardService.widgetContent.update(content => ({
-          ...content,
-          [columnIndex]: event.data,
-        }));
-        this._contentValid.next(event.valid);
-      })
-    );
     this.cdr.detectChanges();
   }
 }
